[Fix] Move authentication controller to ConnectorController.ts  * it applies to all connectors
